+++ ezdxf/entitysection.py
@@ -58,10 +58,23 @@ class AbstractSection(object):
         self._entity_space.write(stream)
         stream.write("  0\nENDSEC\n")
 
+    def create_new_dxf_entity(self, _type, dxfattribs):
+        """ Create new DXF entity add it to th entity database and add it to the entity space.
+        """
+        dxf_entity = self.dxffactory.create_db_entry(_type, dxfattribs)
+        self._entity_space.add_handle(dxf_entity.dxf.handle)
+        return dxf_entity
+
+    def remove_handle(self, handle):
+        self._entity_space.remove(handle)
+
     # start of public interface
 
     def __len__(self):
         return len(self._entity_space)
+
+    def __contains__(self, handle):
+        return handle in self._entity_space
 
     def query(self, query='*'):
         return EntityQuery(iter(self), query)
@@ -90,9 +103,6 @@ class ObjectsSection(ClassesSection):
 
     def roothandle(self):
         return self._entity_space[0]
-
-    def add_handle(self, handle):
-        self._entity_space.add_handle(handle)
 
 
 class EntitySection(AbstractSection):