+++ tests/test_blockssection.py
@@ -11,12 +11,12 @@ from __future__ import unicode_literals
 import unittest
 from io import StringIO
 
+import ezdxf
 from ezdxf.testtools import DrawingProxy, normlines, Tags
-
 from ezdxf.blockssection import BlocksSection
 
 
-class TestBlocksSection(unittest.TestCase):
+class TestBlocksSectionAC1009(unittest.TestCase):
     def setUp(self):
         self.dwg = DrawingProxy('AC1009')
         self.section = BlocksSection(Tags.from_text(TESTBLOCKS), self.dwg)
@@ -60,6 +60,91 @@ class TestBlocksSection(unittest.TestCase):
         self.assertEqual(1, len(entities))  # VERTEX & SEQEND doesn't count
         e = entities[0]
         self.assertEqual(self.dwg, e.drawing)
+
+    def test_delete_block(self):
+        archtick = self.section['_ARCHTICK']
+        entities = list(archtick)
+        archtick_name = archtick.name
+        self.section.delete_block(archtick_name)
+        self.assertTrue(archtick_name not in self.section)
+        db = self.dwg.entitydb
+        self.assertTrue(archtick._block_handle not in db)
+        self.assertTrue(archtick._endblk_handle not in db)
+        for entity in entities:
+            self.assertTrue(entity.dxf.handle not in db)
+
+    def test_delete_all_blocks(self):
+        self.section.delete_all_blocks()
+        blocks = list(self.section)
+        # assure not deleting layout blocks
+        self.assertEqual(2, len(blocks))
+        block_names = [block.name for block in blocks]
+        block_names.sort()
+        self.assertEqual(['$MODEL_SPACE', '$PAPER_SPACE'], block_names)
+
+
+class TestBlocksSectionAC1015(unittest.TestCase):
+    DWG_AC1015 = ezdxf.new('AC1015')  # DXF structure is too complex for dummy constructions.
+
+    def setUp(self):
+        self.dwg = self.DWG_AC1015
+        if 'TestBlock' not in self.dwg.blocks:
+            self.create_test_block()
+
+    def create_test_block(self):
+        block = self.dwg.blocks.new('TestBlock')
+        block.add_line((0, 0), (10, 10))
+        block.add_line((0, 0), (10, 10))
+        block.add_line((0, 0), (10, 10))
+
+    def test_dxf_block_structure(self):
+        self.assertTrue('TestBlock' in self.dwg.blocks)
+        block = self.dwg.blocks['TestBlock']
+        block_record_handle = block.get_block_record_handle()
+
+        # exists an associated block record entry?
+        block_record = self.dwg.sections.tables.block_records.get(block.name)
+        self.assertEqual(block_record_handle, block_record.dxf.handle)
+        self.assertEqual(block_record.dxf.name, block.name)
+
+    def test_delete_block(self):
+        block = self.dwg.blocks['TestBlock']
+        block_name = block.name
+        entities = list(block)
+        block_record_handle = block.get_block_record_handle()
+
+        block_count = len(self.dwg.blocks)
+        self.dwg.blocks.delete_block(block_name)
+
+        # removed from blocks section?
+        self.assertEqual(block_count-1, len(self.dwg.blocks))
+        self.assertTrue(block_name not in self.dwg.blocks)
+
+        # all block related management data deleted?
+        db = self.dwg.entitydb
+        self.assertTrue(block._block_handle not in db)
+        self.assertTrue(block._endblk_handle not in db)
+        self.assertTrue(block_record_handle not in db)
+
+        # removed from block records table?
+        self.assertTrue(block_name not in self.dwg.sections.tables.block_records)
+
+        # all entities deleted ?
+        for entity in entities:
+            self.assertTrue(entity.dxf.handle not in db)
+        # we are done!
+
+    def test_delete_all_blocks(self):
+        self.dwg.blocks.delete_all_blocks()
+        blocks = list(self.dwg.blocks)
+
+        # assure not deleting layout blocks
+        self.assertEqual(2, len(blocks))
+
+        block_names = [block.name for block in blocks]
+        block_names.sort()
+        self.assertEqual(['*Model_Space', '*Paper_Space'], block_names)
+
 
 EMPTYSEC = """  0
 SECTION