+++ NEWS.rst
@@ -6,8 +6,9 @@ Version 0.6.1 - 2014
 
   * Beta status
   * Supported Python versions: CPython 2.7, 3.4 and pypy 2.2.1
-  * NEW: create new layouts
-  * NEW: delete layouts
+  * NEW: create new layouts - Drawing.create_layout(name, dxfattribs=None)
+  * NEW: delete layouts - Drawing.create_layout(name)
+  * NEW: delete blocks - Drawing.blocks.delete_block(name)
   * CHANGE: LWPolyline returns always 5-tuples (x, y, start_width, end_width, bulge). start_width, end_width and bulge
     is 0 if not present.
   * NEW: LWPolyline.get_rstrip_points() -> generates points without appending zeros.